test: move PrometheusObserver tests to their mirror module
The test tree mirrors the source tree one-to-one, but the
PrometheusObserver suites lived under spec/observability while the class
itself lives in node/metrics/spec_observer.py.

Move TestPrometheusObserverUninitialized and TestPrometheusObserverWithRegistry
to a new tests/lean_spec/node/metrics/test_spec_observer.py, carrying the
Prometheus-only fixtures and histogram helper with them.

What stays in spec/observability/test_observer.py now tests only
observer.py: the no-op default, set_observer, and the observe_* context
managers. That file no longer imports the Prometheus adapter or the
metrics registry; the set_observer test uses the local recording
observer instead.

Along the way: add the missing one-line docstrings to every test, give
the uninitialized test a real assertion (the registry stays
uninitialized), and assert the null-observer hooks return None.
